feat: GlobalStyle 도입해 CSS 리셋
