+++ src/App.js
@@ -1,8 +1,10 @@
 import Routers from "./routers/Routers";
+import GlobalStyle from "./styles/GlobalStyle";
 
 const App = (props) => {
   return (
     <>
+      <GlobalStyle />
       <Routers />
     </>
   );
