+++ src/components/class/ClassFrame.jsx
@@ -1,11 +1,11 @@
 import ChatFrame from "./ChatFrame";
-import VideoFrame from "./VideoFrame";
+// import VideoFrame from "./VideoFrame";
 
 const ClassFrame = () => {
   return (
     <>
       <ChatFrame />
-      <VideoFrame />
+      {/* <VideoFrame /> */}
     </>
   );
 };
